reorganized some filed for improved modularity
