+++ src/lfs_tools/guided_learning_activity/student_end/components/ContextComponent.jsx
@@ -1,4 +1,4 @@
-import TextContainerCard from "@/lfs_tools/shared_components/cards/TextContainerCard";
+import TextContainerCard from "@/global_ui_components/cards/TextContainerCard";
 
 const ContextComponent = ({ inquiry }) => {
     return (
